pixel offset refactored to PixelOffsetDecorator
